feat: add kafka_config and kafka_producer for auth-service

+++ services/auth-service/pkg/dto/auth.go
@@ -44,8 +44,8 @@ type RefreshTokenOutput struct {
 }
 
 type TokenRequest struct {
-	UserID     uint
+	UserID     uint64
 	Username   string
 	Role       string
-	PwdVersion int
+	PwdVersion int64
 }
